Feat: #141 - 텍스트 이미지 생성되는 크기 조절
